Fixed bug with the llvm pass dump viewer, split test case from the llvm ir parser code
